requirements.txt
streamlit==1.54.0
plotly==6.5.2
numpy==1.26.4
pandas==2.3.3
scipy==1.17.1
pykrx==1.2.4
tqdm==4.67.3
python-dotenv==1.2.1
anthropic==0.83.0
rich>=13.0
psycopg2-binary>=2.9
